Validated deleteNote & updateNote

+++ NoteTaker/src/main/java/lk/ijse/gdse/springboot/notetaker/controller/NoteController.java
@@ -54,6 +54,9 @@ public class NoteController {
     @PatchMapping(value = "/{noteId}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateNote(@PathVariable("noteId") String noteId,@RequestBody NoteDto note){ //http://localhost:8080/notetaker/api/v1/note/1
         try{
+            if (note == null && (noteId == null || note.equals(""))){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             noteService.updateNote(noteId, note);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (NoteNotFoundException e) {
@@ -65,6 +68,9 @@ public class NoteController {
     @DeleteMapping(value = "/{noteId}")
     public ResponseEntity<Void> deleteNote(@PathVariable("noteId") String noteId){ //http://localhost:8080/notetaker/api/v1/note/1
         try{
+            if (noteId == null || noteId.equals("")){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             noteService.deleteNote(noteId);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (NoteNotFoundException e) {
